feat: make contact name & phone req
additionally, we add ternary operators on contact_name,
contact_phone, and upon_arrival_instructions as these fields
were added after much of the test data was added to dev.
without the ternary operators the fields in the form get set to
'null' (as they don't exist in firebase) and causes crashes
when attempting to update older locations.